code and pom clean up
git-svn-id: https://svn.wso2.org/repos/wso2/carbon/platform/trunk@134126 a5903396-d722-0410-b921-86c7d4935375

+++ components/load-balancer/autoscaler/org.wso2.carbon.mediator.autoscale/src/main/java/org/wso2/carbon/mediator/autoscale/lbautoscale/context/AppDomainContext.java
@@ -145,7 +145,7 @@ public class AppDomainContext extends LoadBalancerContext{
      *
      * @return number of average requests in flight. -1 if there no requests were received
      */
-    public int getAverageRequestsInFlight() {
+    public long getAverageRequestsInFlight() {
         long total = 0;
         for (Integer messageQueueLength : requestTokenListLengths) {
             total += messageQueueLength;
@@ -158,7 +158,7 @@ public class AppDomainContext extends LoadBalancerContext{
         if (log.isDebugEnabled()) {
             log.debug("Total Tokens : "+total+ " : Size: "+size);
         }
-        return (int) total / size;
+        return (long) total / size;
     }
